# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an auto-generated API Docs.",
        "title": "Zocket image process API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "venkateshwarachinnasamy@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/products": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "add a new product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Add Prodcut",
                "operationId": "add-product",
                "parameters": [
                    {
                        "description": "body parameter",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateProduct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "bad request: validate your input params",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "CreateProduct": {
            "type": "object",
            "required": [
                "description",
                "images",
                "name",
                "price"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code is the status code by default, but also can be\na custom code",
                    "type": "integer"
                },
                "data": {
                    "description": "Data accepts any thing as the response data"
                },
                "message": {
                    "description": "Message shows detail thing back to caller",
                    "type": "string"
                },
                "request_id": {
                    "description": "RequestID needs to be used with middleware",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}